% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv_report.R
\name{hiv_report}
\alias{hiv_report}
\title{Produce data visualization of HIV burden}
\usage{
hiv_report(country)
}
\arguments{
\item{country}{Country for which you want to create a PDF file}
}
\description{
`hiv_report` creates a PDF file of
}
\examples{
hiv_report(country = 'Angola')
}
